move func_rotate_frame_local into contact_local.py
