+++ genesis/engine/solvers/rigid/collider/narrowphase_local.py
@@ -16,47 +16,9 @@ import sys
 import gstaichi as ti
 
 import genesis as gs
-from genesis.engine.solvers.rigid.collider import box_contact, contact, diff_gjk, gjk_local, mpr_local, multi_contact_local
+from genesis.engine.solvers.rigid.collider import box_contact, contact, contact_local, diff_gjk, gjk_local, mpr_local, multi_contact_local
 from genesis.engine.solvers.rigid.collider.narrowphase import CCD_ALGORITHM_CODE
 from genesis.utils import array_class, geom_utils as gu
-
-
-@ti.func
-def func_rotate_frame_local(
-    pos: ti.types.vector(3, dtype=gs.ti_float),
-    quat: ti.types.vector(4, dtype=gs.ti_float),
-    contact_pos: ti.types.vector(3, dtype=gs.ti_float),
-    qrot: ti.types.vector(4, dtype=gs.ti_float),
-) -> ti.types.struct(
-    pos=ti.types.vector(3, dtype=gs.ti_float),
-    quat=ti.types.vector(4, dtype=gs.ti_float),
-):
-    """
-    Thread-local version of func_rotate_frame.
-
-    Instead of modifying geoms_state in place, this function takes thread-local
-    pos/quat and returns the updated values.
-
-    Args:
-        pos: Thread-local geometry position
-        quat: Thread-local geometry quaternion
-        contact_pos: Contact position to rotate around
-        qrot: Rotation quaternion for perturbation
-
-    Returns:
-        Struct containing updated pos and quat
-    """
-    # Update quaternion
-    new_quat = gu.ti_transform_quat_by_quat(quat, qrot)
-
-    # Update position
-    rel = contact_pos - pos
-    vec = gu.ti_transform_by_quat(rel, qrot)
-    vec = vec - rel
-    new_pos = pos - vec
-
-    # Return struct with both updated values
-    return ti.Struct(pos=new_pos, quat=new_quat)
 
 
 @ti.func
@@ -181,11 +143,11 @@ def func_convex_convex_contact_local(
                 qrot = gu.ti_rotvec_to_quat(collider_info.mc_perturbation[None] * axis, EPS)
 
                 # Apply perturbation starting from original state
-                ga_result = func_rotate_frame_local(ga_pos_original, ga_quat_original, contact_pos_0, qrot)
+                ga_result = contact_local.func_rotate_frame_local(ga_pos_original, ga_quat_original, contact_pos_0, qrot)
                 ga_pos_current = ga_result.pos
                 ga_quat_current = ga_result.quat
 
-                gb_result = func_rotate_frame_local(
+                gb_result = contact_local.func_rotate_frame_local(
                     gb_pos_original, gb_quat_original, contact_pos_0, gu.ti_inv_quat(qrot)
                 )
                 gb_pos_current = gb_result.pos
